pyproject.toml
[project]
name = "pdfagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.5",
    "langchain-community>=0.3.8",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.2.9",
    "openai>=1.55.0",
    "pdfplumber>=0.11.4",
    "pip>=24.3.1",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "python-fasthtml>=0.10.0",
    "sentence-transformers>=3.3.1",
    "transformers>=4.46.3",
    "uvicorn>=0.32.1",
]
